Moved sources into separate directories.
