+++ tests/nodenorm/test_nodenorm_from_gsheet.py
@@ -1,8 +1,7 @@
-import itertools
 import urllib.parse
 import requests
 import pytest
-from src.babel_validation.sources.google_sheet_test_cases import GoogleSheetTestCases
+from src.babel_validation.sources.google_sheets.google_sheet_test_cases import GoogleSheetTestCases
 
 # We generate a set of tests from the GoogleSheetTestCases.
 gsheet = GoogleSheetTestCases()
